+++ src/css/app.sass
@@ -86,6 +86,9 @@
 .text-no-decoration
   text-decoration: none
 
+.inline-block
+  display: inline-block
+
 q-card.card--light-bg
   background: #f4f0fb
